Places viewport in a subpackage; adds factory function
By viewport into a subpackage, we can limit the interface for how we
interact with it. In this way we can ensure that there's no way to
create a bad viewport with values that don't make mathematical
sense. This acts as a checkpoint and forces any of those bad values to
be revealed, crashing the program as soon as they are detected.